backend/requirements.txt
google-genai>=0.7.0
fastapi>=0.110.0
uvicorn>=0.29.0
google-api-python-client>=2.120.0
google-auth>=2.29.0
requests>=2.31.0
python-docx>=1.1.0
python-multipart>=0.0.9
pydantic>=2.6.0
python-dotenv>=1.0.0
strands-agents[ollama]
duckduckgo-search
openai